+++ yash-env/CHANGELOG.md
@@ -32,7 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
-- The `fcntl_setfl` method from the `System` trait
+- The `fcntl_getfl` and `fcntl_setfl` methods from the `System` trait
 
 ## [0.2.1] - 2024-07-12